Adapt meta-plus to allow running after NEEO-cloud has stopped
This solution makes a number of changes to the m eta0-plus package to support the "After NEEO cloud"situation.
It works together with a minor change to the NEEO-software itself (cp6.js) where most of the "NEEO cloud handling takes place".
All data that "used to be in the NEEO cloud" will not be checked against local storage on the system meta is running on.