Support for snapshot of messages added
metaMessage totally revised.
Unless specified otherwise, this new function now saves a number of messages that is reproduced when a message with loglevel = ERROR is received.

This allows for suppression of all messages, but still receive them when errors are found
logComponents now contains variable produceNrSnapshotWhenError: the number of messages that are saved continuously. Specifying nothing enables 200 messages to be saved, specifying 0 disables this functionality